lib: add translator lib
Add kube-translator.py, which takes two
dictionaries as arguments: one a dict of
templates keyed by name, and the other
a Flux jobspec. The translator matches
the jobspec command to a corresponding
OpenShift template name, and applies
the template parameter overrides
specified in the jobspec's system
attributes. It returns a service,
which is the selected template plus
a dictionary of matched parameter
overrides.